requirements.txt
instabot==0.106.0
python-dotenv==0.10.5
tweepy==3.8.0